Add extra step to enable the Google+ API
If the Google+ API is not enabled for your web application client in the Google Developers Console, you will receive the message: ```Access Not Configured. The API (Google+ API) is not enabled for your project. Please use the Google Developers Console to update your configuration.```

This is the only API needed for `logcabin` to work.